Category Separation Collective Changes
Results are now separated into 3 categories: 'Default', 'Saved', and 'Deleted'.
Users can change the category of an item via the front end interface.

Note: 'Deleted' items are not deleted from the DB, as they still need to be checked against for duplication. They are only 'partially' hidden from in the front end, with the ability to 'restore' them back to a 'Default' state.

Added a created/updated date for an item. An item recieves a created timestamp, and if, when getting new items it find a duplicate in the DB, it only updates its timestamp to match the current date.

Created a crude removeExpired to remove expired items (exact expiry time frame not yet determined)